requirements.txt
streamlit==1.25.0
opencv-python
numpy
Pillow
streamlit-drawable-canvas